+++ packages/db/src/core/utils.ts
@@ -16,7 +16,7 @@ export function getRemoteDatabaseUrl(): string {
 
 export function getAstroStudioUrl(): string {
 	const env = getAstroStudioEnv();
-	return env.ASTRO_STUDIO_URL || 'https://stardate.astro.build';
+	return env.ASTRO_STUDIO_URL || 'https://studio.astro.build';
 }
 
 export function getDbDirectoryUrl(root: URL | string) {
@@ -26,3 +26,24 @@ export function getDbDirectoryUrl(root: URL | string) {
 export function defineDbIntegration(integration: AstroDbIntegration): AstroIntegration {
 	return integration;
 }
+
+/**
+ * Small wrapper around fetch that throws an error if the response is not OK. Allows for custom error handling as well through the onNotOK callback.
+ */
+export async function safeFetch(
+	url: Parameters<typeof fetch>[0],
+	options: Parameters<typeof fetch>[1] = {},
+	onNotOK: (response: Response) => void | Promise<void> = () => {
+		throw new Error(`Request to ${url} returned a non-OK status code.`);
+	}
+): Promise<Response> {
+	const response = await fetch(url, options);
+
+	if (!response.ok) {
+		await onNotOK(response);
+	}
+
+	return response;
+}
+
+export type Result<T> = { success: true; data: T } | { success: false; data: unknown };