+++ packages/db/test/basics.test.js
@@ -13,10 +13,6 @@ describe('astro:db', () => {
 		});
 	});
 
-	// Note (@bholmesdev) generate a random database id on startup.
-	// Ensures database connections don't conflict
-	// when multiple dev servers are run in parallel on the same project.
-	process.env.ASTRO_TEST_RANDOM_DB_ID = 'true';
 	describe('development', () => {
 		let devServer;
 
@@ -26,7 +22,6 @@ describe('astro:db', () => {
 
 		after(async () => {
 			await devServer.stop();
-			process.env.ASTRO_TEST_RANDOM_DB_ID = undefined;
 		});
 
 		it('Prints the list of authors', async () => {
@@ -68,6 +63,14 @@ describe('astro:db', () => {
 
 			const themeDark = $($('.themes-list .theme-dark')[0]).text();
 			expect(themeDark).to.equal('dark mode');
+		});
+
+		it('text fields an be used as references', async () => {
+			const html = await fixture.fetch('/login').then((res) => res.text());
+			const $ = cheerioLoad(html);
+
+			expect($('.session-id').text()).to.equal('12345');
+			expect($('.username').text()).to.equal('Mario');
 		});
 	});
 });